requirements.txt
pytest
playwright
PyHamcrest